refactor: rearange nncf adapter
