+++ otx/algorithms/common/adapters/nncf/utils/__init__.py
@@ -1,0 +1,23 @@
+# Copyright (C) 2022 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+#
+
+from .utils import (
+    is_nncf_enabled,
+    check_nncf_is_enabled,
+    get_nncf_version,
+    load_checkpoint,
+    no_nncf_trace,
+    is_in_nncf_tracing,
+    is_accuracy_aware_training_set,
+)
+
+__all__ = [
+    "is_nncf_enabled",
+    "check_nncf_is_enabled",
+    "get_nncf_version",
+    "load_checkpoint",
+    "no_nncf_trace",
+    "is_in_nncf_tracing",
+    "is_accuracy_aware_training_set",
+]
